backend/requirements.txt
fastapi
uvicorn
openai
python-multipart
google-cloud-texttospeech
